repo-updater: move git.RevList and git.RevListEach from vcs/git to gitserver/commands
