+++ enterprise/internal/rockskip/server_test.go
@@ -19,8 +19,8 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/database/dbtest"
+	"github.com/sourcegraph/sourcegraph/internal/gitserver"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver/gitdomain"
-	"github.com/sourcegraph/sourcegraph/internal/vcs/git"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
 
@@ -247,7 +247,7 @@ func (g SubprocessGit) LogReverseEach(repo string, db database.DB, givenCommit s
 }
 
 func (g SubprocessGit) RevListEach(repo string, db database.DB, givenCommit string, onCommit func(commit string) (shouldContinue bool, err error)) (returnError error) {
-	revList := exec.Command("git", git.RevListArgs(givenCommit)...)
+	revList := exec.Command("git", gitserver.RevListArgs(givenCommit)...)
 	revList.Dir = g.gitDir
 	output, err := revList.StdoutPipe()
 	if err != nil {
@@ -265,7 +265,7 @@ func (g SubprocessGit) RevListEach(repo string, db database.DB, givenCommit stri
 		}
 	}()
 
-	return git.RevListEach(output, onCommit)
+	return gitserver.NewClient(db).RevListEach(output, onCommit)
 }
 
 func (g SubprocessGit) ArchiveEach(repo string, commit string, paths []string, onFile func(path string, contents []byte) error) error {
